merge가 발생해서 일단 branch 에 있는 걸로 남겨두었습니다.

+++ TOHOprojectBusan/Assets/Scripts/Ingame/Enemy.cs
@@ -14,7 +14,7 @@ public class Enemy : MonoBehaviour
         mySequence = seq;
     }
 
-    public void Damage(int atk)
+    public void Damage(float atk)
     {
         HP -= atk;
